+++ system/lib/libc/musl/include/inttypes.h
@@ -22,13 +22,13 @@ uintmax_t strtoumax(const char *__restrict, char **__restrict, int);
 intmax_t wcstoimax(const wchar_t *__restrict, wchar_t **__restrict, int);
 uintmax_t wcstoumax(const wchar_t *__restrict, wchar_t **__restrict, int);
 
-#if UINTPTR_MAX == UINT64_MAX
-#define __PRI64  "l"
-#define __PRIPTR "l"
-#elif defined(__EMSCRIPTEN__)
+#if defined(__EMSCRIPTEN__)
 // Under emscripten __PTRDIFF_TYPE__ and therefor intptr_t are defined to
 // be `long int` even on wasm32.
 #define __PRI64  "ll"
+#define __PRIPTR "l"
+#elif UINTPTR_MAX == UINT64_MAX
+#define __PRI64  "l"
 #define __PRIPTR "l"
 #else
 #define __PRI64  "ll"